+++ src/cjs/interfaces/core.js
@@ -14,6 +14,7 @@
  * @modified 2021-05-31 Added the `drawLib.setConfiguration` function.
  * @modified 2021-05-31 Splitted the large interfaces.ts file into this one and others.
  * @modified 2021-06-21 Added `IBounds.getCenter()`.
+ * @modified 2021-11-16 Added `text` options to the `DrawConfig`.
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 //# sourceMappingURL=core.js.map