+++ src/esm/interfaces/core.js
@@ -13,6 +13,7 @@
  * @modified 2021-05-31 Added the `drawLib.setConfiguration` function.
  * @modified 2021-05-31 Splitted the large interfaces.ts file into this one and others.
  * @modified 2021-06-21 Added `IBounds.getCenter()`.
+ * @modified 2021-11-16 Added `text` options to the `DrawConfig`.
  **/
 export {};
 //# sourceMappingURL=core.js.map